requirements.txt
-e git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
aiohttp==3.5.4
alembic==1.0.8
appdirs==1.4.3
arrow==0.13.1
async-timeout==3.0.1
asyncio==3.4.3
asyncore-wsgi==0.0.9
attrs==19.1.0
bcrypt==3.1.6
beautifulsoup4==4.7.1
bottle==0.12.16
bs4==0.0.1
CacheControl==0.12.5
cachetools==3.1.0
cffi==1.12.2
chardet==3.0.4
chirptext==0.1a18
Click==7.0
colorama==0.4.1
decorator==4.4.0
distlib==0.2.8
distro==1.4.0
dogpile.cache==0.7.1
dominate==2.3.5
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Scss==0.5
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==1.4.0
greenlet==0.4.15
html5lib==1.0.1
idna==2.8
inflect==2.1.0
itsdangerous==1.1.0
jamdict==0.1a4
jaraco.itertools==4.4.2
Jinja2==2.10
libsass==0.18.0
lockfile==0.12.2
lxml==4.3.2
Mako==1.0.8
MarkupSafe==1.1.1
mecab-python3==0.996.1
more-itertools==7.0.0
msgpack==0.6.1
multidict==4.5.2
myougiden==0.8.8
npyscreen==4.10.5
packaging==19.0
pep517==0.5.0
progress==1.5
puchikarui==0.1a1
pycparser==2.19
pyparsing==2.3.1
pyScss==1.3.5
python-dateutil==2.8.0
python-editor==1.0.4
pytoml==0.1.20
PyYAML==5.1
requests==2.21.0
retrying==1.3.3
romkan==0.2.1
six==1.12.0
soupsieve==1.9
SQLAlchemy==1.3.2
SQLAlchemy-Paginator==0.1
SQLAlchemy-Utils==0.33.11
termcolor==1.1.0
urllib3==1.24.1
visitor==0.1.3
web-pdb==1.5.1
webencodings==0.5.1
websockets==6.0
Werkzeug==0.15.0
wikipedia==1.4.0
wolframalpha==3.0.1
wordnik-py3==2.1.2
WTForms==2.2.1
xmltodict==0.12.0
yarl==1.3.0
